pyproject.toml
[project]
name = "winniedapooh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
]

[tool.uv]
package = false
